requirements.txt
asgiref==3.7.2
Django==4.2.2
sqlparse==0.4.4
typing_extensions==4.6.3
flake8==6.0.0
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
djangorestframework==3.14.0
pytz==2023.3
PyYAML==6.0
attrs==23.1.0
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
pyrsistent==0.19.3
djangorestframework-simplejwt==5.2.2
pyjwt==2.7.0
stripe==5.0.0
aiogram==2.25.1
psycopg2-binary==2.9.6
python-dotenv==1.0.0
